Refactoring Large Class using Extract class
